build: group webpack & ignore dist & support yarn1.x
