+++ .lintstagedrc.json
@@ -1,5 +1,5 @@
 {
   "*": [
-      "yarn lint:fix" 
+      "yarn lint:fix"
    ]
 }
