--- src/jison/parser-options-schema.json
@@ -1,13 +1,0 @@
-{
-  "title": "Jison Parser options",
-  "type": "object",
-  "properties": {
-      "token-stack": {
-          "type": "boolean"
-      },
-      "debug": {
-          "type": "boolean"
-      }
-  },
-  "additionalProperties": false
-}
